tests/wpt/tests/css/css-anchor-position/anchor-center-001.html

<!DOCTYPE html>
<title>Tests the anchor-center keyword is parsed and computed as specified</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#anchor-center">
<link rel="author" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="container">
  <div id="target"></div>
</div>

<script>
test_valid_value('align-self', 'anchor-center');
test_valid_value('align-items', 'anchor-center');
test_valid_value('justify-self', 'anchor-center');
test_valid_value('justify-items', 'anchor-center');

test_computed_value('align-self', 'anchor-center');
test_computed_value('align-items', 'anchor-center');
test_computed_value('justify-self', 'anchor-center');
test_computed_value('justify-items', 'anchor-center');
</script>
